#105 refactor: user - 유저 서비스 인터페이스 분리
