+++ user-service/src/main/java/com/fn/eureka/client/userservice/presentation/controller/AuthController.java
@@ -1,7 +1,10 @@
 package com.fn.eureka.client.userservice.presentation.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.fn.common.global.dto.CommonResponse;
 import com.fn.common.global.success.SuccessCode;
